requirements.txt
emails==0.6
python-dotenv==0.17.0